fix: make header responsive, refactors
